Added powerups + textures, no functionality yet. Changes committed:         modified:   Makefile 	modified:   Return-voidCB.cbp.Linux 	modified:   Return-voidCB.depend.Linux 	modified:   Return-voidCB.layout.Linux 	renamed:    Textures/Game/BasicBullet.png -> Textures/Game/Bullets/BasicBullet.png 	renamed:    Textures/Game/PlayerBullet.png -> Textures/Game/Bullets/PlayerBullet.png 	new file:   Textures/Game/Powerups/Bomb.png 	new file:   Textures/Game/Powerups/Money.png 	new file:   Textures/Game/Powerups/Shield.png 	renamed:    Textures/Game/BaseEnemy.png -> Textures/Game/Ships/BaseEnemy.png 	renamed:    Textures/Game/BasicEnemy.png -> Textures/Game/Ships/BasicEnemy.png 	renamed:    Textures/Game/Ship.png -> Textures/Game/Ships/Ship.png 	renamed:    Textures/Game/ShipLeft.png -> Textures/Game/Ships/ShipLeft.png 	renamed:    Textures/Game/ShipRight.png -> Textures/Game/Ships/ShipRight.png 	modified:   include/EnemyBase.h 	modified:   include/Game.h 	modified:   include/Level.h 	modified:   include/Menu.h 	modified:   include/Player.h 	new file:   include/Powerups.h 	modified:   src/Bullet.cpp 	modified:   src/EnemyBase.cpp 	modified:   src/EnemyTypes.cpp 	modified:   src/Game.cpp 	modified:   src/Level.cpp 	modified:   src/Menu.cpp 	modified:   src/Player.cpp 	new file:   src/Powerups.cpp 	modified:   src/main.cpp

+++ include/Player.h
@@ -10,83 +10,82 @@
 #include <GL/freeglut.h>
 #endif
 
-#include <stdio.h> 
+#include <stdio.h>
 #include <cstdlib>
 #include <stdlib.h>
 #include <vector>
 
 #include "Bullet.h"
 
-class Player 
+class Player
 {
     public:
         Player();
         void setup(GLuint *newTextures, float aspectRatio);
         void destroy();
         void respawn(int x, int y);
-        
+
         void takeHealth();
-        
+
         void Tick(float mouseX, float mouseY);
-        void draw();  
+        void draw();
         void fire();
-        
+
         void moveLeft();
         void moveRight();
         void moveUp();
         void moveDown();
-        
+
         void rotateRight();
         void rotateLeft();
 
         void leftImage();
         void rightImage();
         void stationaryImage();
-        
+
         int getHealth();
-        int getWidth();
-        int getHeight();
+        float getWidth();
+        float getHeight();
         float getX();
         float getY();
         bool isAlive();
         bool getVisible();
         bool getInvincible();
-         
+
         int getBulletNum();
-        int getBulletWidth(int i); 
+        int getBulletWidth(int i);
         int getBulletHeight(int i);
         float getBulletX(int i);
-        float getBulletY(int i); 
-        bool getBulletVisible(int i);    
-        
+        float getBulletY(int i);
+        bool getBulletVisible(int i);
+
         void setHealth(int newHealth);
         void setVisible(bool Visible);
         void setBulletVisible(bool visible, int i);
 
-    protected:  
+    protected:
         void flash();
-       
+
         int health;
         int fireRate;
         float angle;
         float width, height;
-        int boundryX, boundryY; 
+        int boundryX, boundryY;
         int invincTimer, flashTimer;
         float aspectRatio, lastMouseX, lastMouseY;
         float directionX, directionY;
 
-        float x, y; 
+        float x, y;
         float speed;
         bool visible, alive, invincible;
         bool drawInvincible;
-        //Bullet bullets[10];
         std::vector<Bullet*> bullets;
         GLuint texture;
         GLuint bulletTexture;
         GLuint PlayerLeftText;
         GLuint PlayerRightText;
         GLuint PlayerText;
-        
+
         void drawShip();
 };
